Remove helix related environment variables and update config struct. Consolidate functions in gql.go to service.go
