+++ internal/config/options.go
@@ -1,7 +1,5 @@
 package config
 
 type Config struct {
-	TwitchHelixClientId     string
-	TwitchHelixSecret string
 	TwitchGQLClientId  string
 }
